+++ src/api/authorization/logout.ts
@@ -1,4 +1,4 @@
-import request from '..';
+import request from '~/api';
 
 const postLogoutApi = async (): Promise<string | false> => {
   try {
